requirements.txt
<<<<<<< HEAD
asttokens==2.4.1
bcrypt==4.2.0
cffi==1.17.1
cryptography==38.0.4
decorator==5.1.1
et-xmlfile==1.1.0
exceptiongroup==1.2.2
executing==2.1.0
gunicorn==23.0.0
ipdb==0.13.13
ipython==8.27.0
jedi==0.19.1
matplotlib-inline==0.1.7
openpyxl==3.1.5
packaging==24.1
parso==0.8.4
pexpect==4.9.0
prompt_toolkit==3.0.47
ptyprocess==0.7.0
pure_eval==0.2.3
pyaes==1.6.1
pycparser==2.22
Pygments==2.18.0
Pyrogram==2.0.106
PySocks==1.7.1
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
stack-data==0.6.3
TgCrypto==1.2.5
thcrypto==0.9.21
thresult==0.9.25
tomli==2.0.1
traitlets==5.14.3
typing_extensions==4.12.2
uvloop==0.20.0
wcwidth==0.2.13
=======
pyrogram==2.0.106
python-dotenv==1.0.1
tgcrypto==1.2.5
uvloop==0.20.0
openpyxl==3.1.5
pytz==2024.1
gunicorn==23.0.0
>>>>>>> refs/remotes/origin/main
